+++ starkinfra/pixstatement/__pixstatement.py
@@ -6,16 +6,16 @@ from ..utils.checks import check_date, check_datetime
 class PixStatement(Resource):
     """# PixStatement object
     The PixStatement object stores information about all the transactions that happened on
-    a specific day at the workspace. It must be created by the used before it can be
+    a specific day at the workspace. It must be created by the user before it can be
     accessed by the user. This feature is only available for direct participants.
     ## Parameters (required):
     - after [datetime.date]: transactions that happened at this date are stored in the PixStatement , must be the same as before. ex: (2022-01-01)
     - before [datetime.date]: transactions that happened at this date are stored in the PixStatement, must be the same as after. ex: (2022-01-01)
-    - statement_type [string, default None]: statement type. Options: ["interchange", "interchangeTotal", "transaction"]
+    - statement_type [string, default None]: types of entities to include in statement. Options: ["interchange", "interchangeTotal", "transaction"]
     ## Attributes (return-only):
     - id [string, default None]: unique id returned when the PixStatement is created. ex: "5656565656565656"
     - status [string, default None]: current PixStatement status. ex: "success" or "failed"
-    - transaction_count [integer]: number of transactions that happened during the day that the PixStatement was requested. ex 11
+    - transaction_count [integer, default None]: number of transactions that happened during the day that the PixStatement was requested. ex 11
     - created [datetime.datetime, default None]: creation datetime for the PixStatement. ex: datetime.datetime(2020, 3, 10, 10, 30, 0, 0)
     - updated [datetime.datetime, default None]: latest update datetime for the PixStatement. ex: datetime.datetime(2020, 3, 10, 10, 30, 0, 0)
     """
@@ -115,8 +115,8 @@ def page(cursor=None, limit=None, after=None, before=None, status=None, ids=None
 
 
 def csv(id, user=None):
-    """# Retrieve a specific PixStatement csv file by its id.
-    Receive a single PixStatement csv file generated in the Stark Infra API by its id.
+    """# Retrieve a .cvs PixStatement
+    Retrieve a specific PixStatement by its ID in a .csv file.
     ## Parameters (required):
     - id [string]: object unique id. ex: "5656565656565656"
     ## Parameters (optional):